% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.coord.R
\name{heatmap.coord}
\alias{heatmap.coord}
\title{Logical Dataframe Coordinates Heatmap}
\usage{
heatmap.coord(data, bool = NULL, mode = "isnot.na")
}
\arguments{
\item{data}{\code{data.frame} in whose coordinates are being looking for.}

\item{bool}{Simple boolean expression, given as a character string, first the
operator, then the number.}

\item{mode}{Only needed when no boolean expression is given in \code{bool}.
\code{"is.na"} \code{mode} is for searching for NA values in \code{data}, and
\code{"isnot.na"} \code{mode} is for searching for not NA in \code{data}.}
}
\value{
A plot generated by \code{\link[NMF]{aheatmap}} showing in red the
coordinates of the found matches.
}
\description{
Perform the logical searching of \code{\link{which.coord}}, and plot the
results as a NMF package based heatmap.
}
