% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.anova.R
\name{multiple.anova}
\alias{multiple.anova}
\title{Multiple Analysis of Variance}
\usage{
multiple.anova(data, group.var, vars = colnames(data))
}
\arguments{
\item{data}{data.frame containing the values of the variables for each
sample. Values of the variables selected in `vars` must be `numeric`.}

\item{group.var}{Vector containing group identifiers. Its length must match
with the number of `data` rows, and its elements must be in the same order.}

\item{vars}{String vector containing the names of the variables
(matching colnames(data)) to analyze. Values of the selected `vars` must be
`numeric`.}
}
\value{
Vector containing resulting p-values for each test for each variable.
}
\description{
Perform ANOVA separately for each of the variables in a
`data.frame`.
}
