% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.beta.R
\name{permanova.beta}
\alias{permanova.beta}
\title{PERMANOVA analysis}
\usage{
permanova.beta(
  physeq,
  method = "bray",
  variables,
  permutations = 999,
  repetitions = 1,
  print.results = TRUE,
  show.plot = TRUE,
  filename = NULL
)
}
\arguments{
\item{physeq}{A phyloseq object from package \pkg{phyloseq}.}

\item{method}{Short string indicating the distance measure to take into account for beta
diversity calculation. It has to be the same as the accepted strings for the
function \code{\link[phyloseq]{distance}}. For example:
\itemize{
  \item \code{bray}: Bray-Curtis
  \item \code{jaccard}: Jaccard.
  \item \code{unifrac}: Original (unweighted) UniFrac distance.
  \item \code{wunifrac}: Weighted-UniFrac distance.
  \item ...
}}

\item{variables}{Vector of strings containing the name, always in first position, of the
metadata variable to analyse. Following variables will be included in the
analysis as confounders.}

\item{permutations}{Number of permutations for PERMANOVA analysis (\code{\link[vegan]{adonis2}}).}

\item{repetitions}{Number of times that PERMANOVA analysis will be performed.}

\item{print.results}{If \code{TRUE}, results of the PERMANOVA analysis will be printed in console.}

\item{show.plot}{If \code{TRUE}, PCoA plot of the beta diversity analysis will be ploted in Rstudio.}

\item{filename}{If \code{NULL}, no plot will be saved. Otherwise, provide the string name of the file
that will contain the PCoA plot of the beta diversity analysis.}
}
\description{
It performs a PERMANOVA analysis through \code{\link[vegan]{adonis2}} multiple times and
it takes the mean of the all resulting p-value as a more robust method than
doing just 1 iteration of PERMANOVA.
}
