% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ggboxplot.R
\name{plot.ggboxplot}
\alias{plot.ggboxplot}
\title{Dichotomic Boxplot with t.test}
\usage{
\method{plot}{ggboxplot}(data, y, x, facet = NULL)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing data.}

\item{y}{\code{character} string. The name of the \code{numeric} variable to plot whose
distribution will be shown through boxplots.}

\item{x}{\code{character} string. The name of the categoric variable determining groups.
In case of dichotomic variable, a t test will be performed.}

\item{facet}{\code{character} string, optional.
Variable chosen for divide boxplots through different categories.}
}
\value{
A \pkg{ggplot2} object representing boxplots of different groups.
A t test p.value is shown when group number is 2.
}
\description{
Dichotomic Boxplot with t.test
}
