% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presence.grouping.R
\name{presence.grouping}
\alias{presence.grouping}
\title{Presence/Absence Grouping}
\usage{
presence.grouping(data, vars = colnames(data))
}
\arguments{
\item{data}{\code{data.frame} with samples in rows and variables in columns.}

\item{vars}{Variables from \code{data} to be classified. All of them must be \code{numeric}.}
}
\value{
A nested list. Each element (variable) contains a character vector
of the names of the samples from each group, presence and absence, respectively,
and \code{NA} if applicable.
}
\description{
Takes a \code{data.frame} and classifies each \code{numeric}
variable into two sample groups: presence (unequal to zero) and absence
(equal to zero).
}
