% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.cor.R
\name{view.cor}
\alias{view.cor}
\title{Filter and View Correlation Assessment}
\usage{
view.cor(x, y, threshold = 0.5, method = "pearson")
}
\arguments{
\item{x, y}{\code{data.frames} with the first and the second sets of variables,
respectively, to be compared each other. Variables must be in columns.
Observations must be in rows, and in the same order in both sets.}

\item{threshold}{Minimum threshold value for filtering correlation results.
It applies to positive and negative correlations.}

\item{method}{Correlation measure to assess. Options are the same as for
\code{\link[stats]{cor}}. Default = \code{"pearson"}.}
}
\value{
\code{numeric} vector with the results of the assessed correlations.
}
\description{
This function outputs the correlation results of multiple
variables as a \code{numeric} vector, and can be filtered by a minimum threshold.
}
