% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.silhouette.R
\name{kmeans.silhouette}
\alias{kmeans.silhouette}
\title{Silhouette Method for K-means Clustering}
\usage{
kmeans.silhouette(data, print.result = FALSE, print.plot = FALSE)
}
\arguments{
\item{data}{\code{data.frame} with samples in rows and variables in columns.}

\item{print.result}{Default = \code{FALSE}. If \code{TRUE}, the optimal number of clusters
and its Silhouette score is printed as a message in the R terminal.}

\item{print.plot}{Default = \code{FALSE}. If \code{TRUE}, a simple scatter plot will be
generated showing the silhouette scores for each potential number of
clusters.}
}
\value{
\code{numeric} value. Number of optimal clusters.
}
\description{
This function performs Silhouette method in order to assess the
optimal number of clusters for given data, when clustering by K-means
algorithm.
}
