% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.anova.R
\name{multiple.anova}
\alias{multiple.anova}
\title{Multiple Analysis of Variance}
\usage{
multiple.anova(data, group.var, vars = colnames(data))
}
\arguments{
\item{data}{\code{data.frame} containing the values of the variables for each
sample. Values of the variables selected in \code{vars} must be \code{numeric}.}

\item{group.var}{Vector containing group identifiers. Its length must match
with the number of \code{data} rows, and its elements must be in the same order.}

\item{vars}{Default = \code{colnames(data)}. String vector containing the names of the variables
(matching \code{colnames(data)}) to analyze. Values of the selected \code{vars} must be
\code{numeric}.}
}
\value{
Vector containing resulting p-values for each test for each variable.
}
\description{
Perform ANOVA separately for each of the variables in a
\code{data.frame}.
}
