% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presence.grouping.R
\name{presence.grouping}
\alias{presence.grouping}
\title{Presence/Absence Grouping}
\usage{
presence.grouping(data, vars = colnames(data), lower.limit = 2)
}
\arguments{
\item{data}{\code{data.frame} with samples in rows and variables in columns.}

\item{vars}{Variables from \code{data} to be classified. All of them must be \code{numeric}.}

\item{lower.limit}{Minimum number allowed of members in each group. Minimum for t test is 2.}
}
\value{
A nested list. Each element (variable) contains a character vector
of the names of the samples from each group, presence and absence, respectively,
and \code{NA} if applicable.
}
\description{
Takes a \code{data.frame} and classifies each \code{numeric}
variable into two sample groups: presence (unequal to zero) and absence
(equal to zero).

Originally, this function was created as a useful precedent to
\code{\link{multiple.t.test}} function.
}
