% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.cor.R
\name{view.cor}
\alias{view.cor}
\title{Filter and View Correlation Assessment}
\usage{
view.cor(vars1, vars2, threshold = 0.5, method = "pearson")
}
\arguments{
\item{vars1}{`data.frame` with the first set of variables, in columns, to
compare with the second set from `vars2`. Observations are in rows, and they
must match in order with those in `vars2`.}

\item{vars2}{`data.frame` with the second set of variables, in columns, to
compare with the first set from `vars1`. Observations are in rows, and they
must match in order with those in `vars1`.}
}
\value{
`numeric` vector with the results of the assessed correlations.
}
\description{
This function outputs the correlation results of multiple
variables as a `numeric` vector, and can be filtered by a minimum threshold.
}
